Refactor to a shared hook for hidden field state
